Karl Hegbloom's fixes to suppress compiler warnings on amd64.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/branches/gnucash-gnome2-dev@11564 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/backend/postgres/putil.h
@@ -318,7 +318,7 @@ int finishQuery(PGBackend *be);
 #define COMP_INT64(sqlname,fun,ndiffs) { 			\
    if (strtoll (DB_GET_VAL(sqlname,0), NULL, 0) != fun) {	\
       PINFO("mis-match: %s sql='%s', eng='%lld'", sqlname, 	\
-         DB_GET_VAL (sqlname,0), fun); 				\
+	 DB_GET_VAL (sqlname,0), (long long int) fun);	\
       ndiffs++; 						\
    }								\
 }
